+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/accountinfo/AccountInfo.java
@@ -11,7 +11,6 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import lombok.Builder;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 import lombok.Singular;
 import lombok.extern.jackson.Jacksonized;
 
@@ -38,6 +37,15 @@ public final class AccountInfo {
      * The amount of CCD owned by this account.
      */
     private final CCDAmount accountAmount;
+    /**
+     * The available (unencrypted) balance of the account (i.e. that can be transferred
+     * or used to pay for transactions). This is the balance ({@link AccountInfo#accountAmount})
+     * minus the locked amount.
+     * The locked amount is the maximum of the amount in the release schedule ({@link AccountInfo#accountReleaseSchedule})
+     * and the total amount that is actively staked or in cooldown (inactive stake, {@link AccountInfo#cooldowns}).
+     * This was introduced in node version 7.0.
+     */
+    private final CCDAmount availableBalance;
     /**
      * The account threshold for this account i.e., how
      * many credentials that needs to sign transactions for this account.